@@ -1,27 +1,41 @@
 {
   "nativescript": {
-    "id": "net.tangrainc.gridviewsample",
+    "id": "com.tangrainc.gridviewsample",
     "tns-ios": {
-      "version": "2.5.0"
+      "version": "3.0.0-rc.1"
     },
     "tns-android": {
-      "version": "2.5.0"
+      "version": "3.0.0-rc.1"
     }
   },
   "scripts": {
-    "android": "npm uninstall nativescript-grid-view && tns run android",
-    "ios": "npm uninstall nativescript-grid-view && tns run ios --emulator"
+    "android": "npm uninstall nativescript-grid-view && tns platform remove android && tns platform add android@rc && tns run android --emulator",
+    "ios": "npm uninstall nativescript-grid-view && tns run ios --emulator",
+    "ns-bundle": "ns-bundle",
+    "start-android-bundle": "npm run ns-bundle --android --start-app --uglify",
+    "start-ios-bundle": "npm run ns-bundle --ios --start-app --uglify",
+    "build-android-bundle": "npm run ns-bundle --android --build-app",
+    "build-ios-bundle": "npm run ns-bundle --ios --build-app"
   },
   "dependencies": {
     "nativescript-grid-view": "file:../bin/dist",
-    "tns-core-modules": "^2.5.0"
+    "tns-core-modules": "rc"
   },
   "devDependencies": {
+    "awesome-typescript-loader": "~3.0.0-beta.9",
     "babel-traverse": "6.21.0",
     "babel-types": "6.21.0",
     "babylon": "6.14.1",
+    "copy-webpack-plugin": "~3.0.1",
+    "extract-text-webpack-plugin": "~2.0.0-beta.4",
     "lazy": "1.0.11",
-    "nativescript-dev-typescript": "0.3.5",
-    "typescript": "2.1.4"
+    "nativescript-css-loader": "~0.26.0",
+    "nativescript-dev-typescript": "0.4.2",
+    "nativescript-dev-webpack": "^0.3.7",
+    "raw-loader": "~0.5.1",
+    "resolve-url-loader": "~1.6.0",
+    "typescript": "2.2.2",
+    "webpack": "2.2.0",
+    "webpack-sources": "~0.1.3"
   }
 }
